Review comments: configurable batch size, make new inner classes top level classes
